Move data moving code to bootstrap.go
